+++ ui/mods/com.pa.quitch.qaipersonalities/new_game.js
@@ -7,77 +7,114 @@ function qaipersonalities() {
 
   qaipersonalitiesLoaded = true;
 
-  api.debug.log("Adding Quitch AI Personalities");
+  api.debug.log("Adding AI Personalities for Skirmish and Multiplayer");
 
   var aiPersonalities = model.aiPersonalities();
 
   var newPersonalities = {
-    "AIP-Absurd Aggressive": {
-      display_name: "!LOC:Absurd Aggressive",
+    "AIP-Aggressive": {
+      display_name: "!LOC:AIP-Aggressive",
       neural_data_mod: 1.33,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Air": {
-      display_name: "!LOC:Absurd Air",
+    "AIP-Air": {
+      display_name: "!LOC:AIP-Air",
       percent_vehicle: 0,
       percent_bot: 0,
       percent_air: 1,
       percent_naval: 0,
       percent_orbital: 0,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Bot": {
-      display_name: "!LOC:Absurd Bot",
+    "AIP-Bot": {
+      display_name: "!LOC:AIP-Bot",
       percent_vehicle: 0,
       percent_bot: 0.95,
       percent_air: 0,
       percent_naval: 0,
       percent_orbital: 0.05,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Cautious": {
-      display_name: "!LOC:Absurd Cautious",
+    "AIP-Cautious": {
+      display_name: "!LOC:AIP-Cautious",
       neural_data_mod: 0.75,
       min_basic_fabbers: 4,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Fast Tech": {
-      display_name: "!LOC:Absurd Fast Tech",
+    "AIP-Defender": {
+      display_name: "!LOC:AIP-Defender",
+      metal_drain_check: 0.71,
+      energy_drain_check: 0.8,
+      metal_demand_check: 0.54,
+      energy_demand_check: 0.65,
+      adv_eco_mod: 1,
+      min_basic_fabbers: 4,
+      personality_tags: ["Vanilla", "PreventsWaste"],
+    },
+    "AIP-Economist": {
+      display_name: "!LOC:AIP-Economist",
+      metal_drain_check: 0.71,
+      energy_drain_check: 0.8,
+      adv_eco_mod: 1,
+      min_basic_fabbers: 4,
+      personality_tags: ["Vanilla", "PreventsWaste"],
+    },
+    "AIP-Fast Tech": {
+      display_name: "!LOC:AIP-Fast Tech",
       adv_eco_mod: 0,
       adv_eco_mod_alone: 0,
       min_basic_fabbers: 4,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Galactic War": {
-      display_name: "!LOC:Absurd Galactic War",
+    "AIP-Foundation": {
+      display_name: "!LOC:AIP-Foundation",
       fabber_to_factory_ratio_basic: 1.5,
-      fabber_to_factory_ratio_advanced: 1.0,
       min_basic_fabbers: 1,
       max_basic_fabbers: 15,
       max_advanced_fabbers: 50,
       personality_tags: ["Vanilla", "PreventsWaste"],
-      metalDrainCheck: 0.45,
-      metalDemandCheck: 0.562,
-      energyDrainCheck: 0.62,
-      energyDemandCheck: 0.712
+      percent_land: 0.05,
+      percent_orbital: 0.05,
+      percent_air: 0.55,
+      percent_naval: 0.35,
+      metal_drain_check: 0.75,
+      energy_drain_check: 0.85,
+      metal_demand_check: 0.75,
+      energy_demand_check: 0.85,
     },
-    "AIP-Absurd Land": {
-      display_name: "!LOC:Absurd Land",
+    "AIP-Land": {
+      display_name: "!LOC:AIP-Land",
       percent_vehicle: 0.6,
       percent_bot: 0.4,
       percent_air: 0,
       percent_naval: 0,
       percent_orbital: 0,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Low Tech": {
-      display_name: "!LOC:Absurd Low Tech",
+    "AIP-Legonis Machina": {
+      display_name: "!LOC:AIP-Legonis Machina",
+      fabber_to_factory_ratio_basic: 1.5,
+      min_basic_fabbers: 1,
+      max_basic_fabbers: 15,
+      max_advanced_fabbers: 50,
+      personality_tags: ["Vanilla", "PreventsWaste"],
+      percent_land: 0.55,
+      percent_air: 0.35,
+      percent_naval: 0.05,
+      percent_orbital: 0.05,
+      metal_drain_check: 0.75,
+      energy_drain_check: 0.85,
+      metal_demand_check: 0.75,
+      energy_demand_check: 0.85,
+    },
+    "AIP-Low Tech": {
+      display_name: "!LOC:AIP-Low Tech",
       adv_eco_mod: 3,
       adv_eco_mod_alone: 2,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Naval": {
-      display_name: "!LOC:Absurd Naval",
+    "AIP-Naval": {
+      display_name: "!LOC:AIP-Naval",
       percent_open_vehicle: 0,
       percent_open_bot: 0,
       percent_open_air: 1,
@@ -88,19 +125,35 @@ function qaipersonalities() {
       percent_air: 0.2,
       percent_naval: 0.8,
       percent_orbital: 0,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Orbital": {
-      display_name: "!LOC:Absurd Orbital",
+    "AIP-Orbital": {
+      display_name: "!LOC:AIP-Orbital",
       percent_vehicle: 0.1125,
       percent_bot: 0.0625,
       percent_air: 0.05,
       percent_naval: 0,
       percent_orbital: 0.775,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Rush": {
-      display_name: "!LOC:Absurd Rush",
+    "AIP-Revenants": {
+      display_name: "!LOC:AIP-Revenants",
+      fabber_to_factory_ratio_basic: 1.5,
+      min_basic_fabbers: 1,
+      max_basic_fabbers: 15,
+      max_advanced_fabbers: 50,
+      personality_tags: ["Vanilla", "PreventsWaste"],
+      percent_land: 0.15,
+      percent_air: 0.15,
+      percent_naval: 0.1,
+      percent_orbital: 0.6,
+      metal_drain_check: 0.75,
+      energy_drain_check: 0.85,
+      metal_demand_check: 0.75,
+      energy_demand_check: 0.85,
+    },
+    "AIP-Rush": {
+      display_name: "!LOC:AIP-Rush",
       percent_bot: 1,
       percent_vehicle: 0,
       percent_air: 0,
@@ -108,35 +161,61 @@ function qaipersonalities() {
       percent_orbital: 0,
       neural_data_mod: 1.33,
       adv_eco_mod: 3,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Tank": {
-      display_name: "!LOC:Absurd Tank",
+    "AIP-Swarm": {
+      display_name: "!LOC:AIP-Swarm",
+      metal_demand_check: 0.99,
+      energy_demand_check: 0.99,
+      min_advanced_fabbers: 1,
+      personality_tags: ["Vanilla", "PreventsWaste"],
+    },
+    "AIP-Synchronous": {
+      display_name: "!LOC:AIP-Synchronous",
+      fabber_to_factory_ratio_basic: 1.5,
+      min_basic_fabbers: 1,
+      max_basic_fabbers: 15,
+      max_advanced_fabbers: 50,
+      personality_tags: ["Vanilla", "PreventsWaste"],
+      percent_land: 0.3,
+      percent_air: 0.3,
+      percent_naval: 0.05,
+      percent_orbital: 0.35,
+      metal_drain_check: 0.75,
+      energy_drain_check: 0.85,
+      metal_demand_check: 0.75,
+      energy_demand_check: 0.85,
+    },
+    "AIP-Tank": {
+      display_name: "!LOC:AIP-Tank",
       percent_vehicle: 0.95,
       percent_bot: 0,
       percent_air: 0,
       percent_naval: 0,
       percent_orbital: 0.05,
-      personality_tags: ["Vanilla", "PreventsWaste"]
+      personality_tags: ["Vanilla", "PreventsWaste"],
     },
-    "AIP-Absurd Turtle": {
-      display_name: "!LOC:Absurd Turtle",
+    "AIP-Turtle": {
+      display_name: "!LOC:AIP-Turtle",
+      metal_demand_check: 0.54,
+      energy_demand_check: 0.65,
       neural_data_mod: 0.25,
-      adv_eco_mod: 0,
+      adv_eco_mod: 0.5,
       adv_eco_mod_alone: 0,
       fabber_to_factory_ratio_basic: 5,
       fabber_to_factory_ratio_advanced: 5,
       fabber_alone_on_planet_mod: 3,
       min_basic_fabbers: 5,
-      max_basic_fabbers: 60,
-      max_advanced_fabbers: 60,
-      personality_tags: ["Vanilla", "PreventsWaste"]
-    }
+      personality_tags: ["Vanilla", "PreventsWaste"],
+    },
   };
 
   var baseline = aiPersonalities.Absurd;
 
-  newPersonalities = _.mapValues(newPersonalities, function(personality, name) {
+  newPersonalities = _.mapValues(newPersonalities, function (
+    personality,
+    name
+  ) {
     var result = _.extend(_.clone(baseline), personality);
     result["name"] = name;
     return result;